pyproject.toml
[tool.poetry]
name = "player_wot_spy"
version = "0.1.0"
description = ""
authors = ["levmorgot <levmorgot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector = "^2.2.9"
requests = "^2.26.0"
asyncio = "^3.4.3"
kafka-python = "^2.0.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
